Split user list UI into focused components
